+++ integration-tests/emulated/chains/parachains/bulletin/bulletin-polkadot/src/genesis.rs
@@ -1,0 +1,77 @@
+// Copyright (C) Parity Technologies and the various Polkadot contributors, see Contributions.md
+// for a list of specific contributors.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Substrate
+use sp_core::storage::Storage;
+
+// Cumulus
+use bulletin_polkadot_runtime::xcm_config::TreasuryAccount;
+use emulated_integration_tests_common::{
+	accounts, build_genesis_storage, collators, SAFE_XCM_VERSION,
+};
+use parachains_common::Balance;
+use polkadot_runtime_constants::currency::UNITS as DOT;
+
+pub const PARA_ID: u32 = 1010;
+pub const ENDOWMENT: u128 = 1_000 * DOT;
+pub const ED: Balance = bulletin_polkadot_runtime::ExistentialDeposit::get();
+
+pub fn genesis() -> Storage {
+	let genesis_config = bulletin_polkadot_runtime::RuntimeGenesisConfig {
+		system: bulletin_polkadot_runtime::SystemConfig::default(),
+		balances: bulletin_polkadot_runtime::BalancesConfig {
+			balances: accounts::init_balances()
+				.into_iter()
+				.chain([TreasuryAccount::get()])
+				.map(|k| (k, ENDOWMENT))
+				.collect(),
+			dev_accounts: None,
+		},
+		parachain_info: bulletin_polkadot_runtime::ParachainInfoConfig {
+			parachain_id: PARA_ID.into(),
+			..Default::default()
+		},
+		collator_selection: bulletin_polkadot_runtime::CollatorSelectionConfig {
+			invulnerables: collators::invulnerables().iter().cloned().map(|(acc, _)| acc).collect(),
+			candidacy_bond: ED * 16,
+			..Default::default()
+		},
+		session: bulletin_polkadot_runtime::SessionConfig {
+			keys: collators::invulnerables()
+				.into_iter()
+				.map(|(acc, aura)| {
+					(
+						acc.clone(),                                     // account id
+						acc,                                             // validator id
+						bulletin_polkadot_runtime::SessionKeys { aura }, // session keys
+					)
+				})
+				.collect(),
+			..Default::default()
+		},
+		polkadot_xcm: bulletin_polkadot_runtime::PolkadotXcmConfig {
+			safe_xcm_version: Some(SAFE_XCM_VERSION),
+			..Default::default()
+		},
+		..Default::default()
+	};
+
+	build_genesis_storage(
+		&genesis_config,
+		bulletin_polkadot_runtime::WASM_BINARY
+			.expect("WASM binary was not built, please build it!"),
+	)
+}